Move tests for problems 1.8 and 1.9 to combined chapter test.
